refactor shared-ui index exports
