+++ src/shared/ui/datePicker/lib/helpers/getInputValueDate.ts
@@ -1,4 +1,4 @@
-import { getInputValueFromDate } from '@/shared/ui/datePicker/lib';
+import { getInputValueFromDate } from '../../lib/utils';
 
 export const getInputValueDate = (value: Date): string => {
   return `${getInputValueFromDate(value)}`;
