--- src/shared/ui/datePicker/lib/helpers/index.ts
@@ -1,3 +1,0 @@
-export * from './addNumberDay';
-export * from './getInputValueDate';
-export * from './updateValueOnPopupCloseAction';
