--- src/shared/ui/datePicker/lib/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useInputValueRange';
-export * from './useShowPopup';
