--- src/shared/ui/datePickerMultiple/lib/helpers/index.ts
@@ -1,3 +1,0 @@
-export * from './addNumberDayMultiple';
-export * from './getInputValueDateMultiple';
-export * from './updateValueOnPopupCloseActionMultiple';
