--- src/shared/ui/datePickerMultiple/lib/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useInputValueRangeMultiple';
-export * from './useShowPopupMultiple';
