--- src/shared/ui/datePickerMultiple/lib/index.ts
@@ -1,3 +1,0 @@
-export * from './helpers';
-export * from './hooks';
-export * from './utils';
