+++ src/shared/ui/index.ts
@@ -1,30 +1,36 @@
-export * from './Loading/Loading';
-export * from './backBtn/BackBtn';
-export * from './button/Button';
-export * from './card/Card';
-export * from './chart/Chart';
-export * from './checkbox/Checkbox';
-export * from './customToast/CustomToast';
-export * from './datePicker/ui/DatePicker';
-export * from './datePickerMultiple/ui/DatePickerMultiple';
-export * from './devMode/DevMode';
-export * from './forms/FormCheckbox';
+export { Loading } from './Loading/Loading';
+export { BackBtn } from './backBtn/BackBtn';
+export { Button } from './button/Button';
+export { Card } from './card/Card';
+export { Chart } from './chart/Chart';
+export type { ChartData } from './chart/Chart';
+export { Checkbox } from './checkbox/Checkbox';
+export type { CheckboxPropsProps } from './checkbox/Checkbox';
+export { CustomToast } from './customToast/CustomToast';
+export type { CustomToastProps } from './customToast/CustomToast';
+export { DatePicker } from './datePicker/ui/DatePicker';
+export { DatePickerMultiple } from './datePickerMultiple/ui/DatePickerMultiple';
+export { DevMode } from './devMode/DevMode';
+export { FormCheckbox } from './forms/FormCheckbox';
 export { FormTextfieldArea } from './forms/FormTextFieldArea';
-export * from './forms/FormTextfield';
-export * from './input/Input';
-export * from './mobileMenu/MobileMenu';
-export * from './modal/Modal';
-export * from './pagination/lib/paginationLogic';
-export * from './pagination/ui/PaginationWithSelect';
-export * from './pagination/ui/paginationInitial/Pagination';
-export * from './pagination/ui/paginationInitial/pageLink/PageLink';
-export * from './photoProfile/PhotoProfile';
-export * from './photosSwiper/PhotosSwiper';
-export * from './post/Post';
-export * from './privacy/Privacy';
-export * from './radio/Radio';
-export * from './select/Select';
-export * from './tabs/TabSwitcher';
-export * from './textArea/TextArea';
-export * from './typography/Typography';
-export * from './wrapper/Wrapper';
+export { FormTextfield } from './forms/FormTextfield';
+export { Input } from './input/Input';
+export type { InputProps } from './input/Input';
+export { MobileMenu } from './mobileMenu/MobileMenu';
+export { Modal } from './modal/Modal';
+export type { ModalProps } from './modal/Modal';
+export { usePagination } from './pagination/lib/paginationLogic';
+export { PaginationWithSelect } from './pagination/ui/PaginationWithSelect';
+export type { SelectOptionsType } from './pagination/ui/PaginationWithSelect';
+export { Pagination } from './pagination/ui/paginationInitial/Pagination';
+export { PhotoProfile } from './photoProfile/PhotoProfile';
+export { PhotosSwiper } from './photosSwiper/PhotosSwiper';
+export { Post } from './post/Post';
+export { Privacy } from './privacy/Privacy';
+export { Radio } from './radio/Radio';
+export { SelectUI } from './select/Select';
+export { TabSwitcher } from './tabs/TabSwitcher';
+export { TextArea } from './textArea/TextArea';
+export type { TextAreaProps } from './textArea/TextArea';
+export { Typography } from './typography/Typography';
+export { Wrapper } from './wrapper/Wrapper';
