+++ src/shared/ui/modal/Modal.tsx
@@ -7,7 +7,7 @@ import { clsx } from 'clsx';
 
 import s from './Modal.module.scss';
 
-export type Props = {
+export type ModalProps = {
   className?: string;
   onOpenChange: (value: boolean) => void;
   open: boolean;
@@ -15,7 +15,7 @@ export type Props = {
   title?: string;
 } & Omit<ComponentPropsWithoutRef<typeof Dialog.Dialog>, 'onOpenChange' | 'open'>;
 
-export const Modal = ({ children, className, style, title, ...props }: Props) => (
+export const Modal = ({ children, className, style, title, ...props }: ModalProps) => (
   <Dialog.Root {...props}>
     <Dialog.Portal>
       <Dialog.Overlay className={s.DialogOverlay} />
