+++ src/shared/ui/pagination/ui/paginationInitial/Pagination.tsx
@@ -1,8 +1,10 @@
 import SvgArrowIosBack from '@/../public/assets/components/ArrowIosBack';
 import SvgArrowIosForward from '@/../public/assets/components/ArrowIosForward';
-import { PageLink, usePagination } from '@/shared/ui';
 
 import s from './Pagination.module.scss';
+
+import { usePagination } from '../../lib/paginationLogic';
+import { PageLink } from './pageLink/PageLink';
 
 type Props = {
   currentPage: number;
