+++ src/widget/modals/deletePhotoModal/DeletePhotoModal.module.scss
@@ -2,7 +2,7 @@
   display: flex;
   gap: 12px;
   align-items: center;
-  justify-content: end;
+  justify-content: flex-end;
 
   width: 100%;
   margin-top: 54px;
